Add SharedSeed and SeedableRandomState.
